readme: document installation and contributing
Previously the documentation regarding installation was
lacking some useful information. Though surely enough for a user who
knows their way around in Firefox, it's not enough for someone who
just "wants it running".

Secondly, the contributing guide(tm) was one sentence; this commit
expands it a little, explaining the commit message format as well as
what is and what is not likely to get merged.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>