theme: ensure a.btn elements behave like buttons
Ensure a.btn elements behave like buttons and not links, by modifying
the relevant selectors to exclude any `a`s that have the `btn` class.

That way, anchors displayed as buttons behave in the expected way,
without any ugly and weird artefacts that would otherwise occur (but are
expected for anchors).

Signed-off-by: Luca Zeuch <l-zeuch@email.de>