css: add CodeMirror styling
Prepare for the upcoming CodeMirror addition to the CC edit page in
light of botlabs-gg/yagpdb#1656.

The CSS is blatantly copied from https://codemirror.net/5/theme/ with a
few removals to make it work with the implementation.

If you're reading this and use this theme do let me know of any visual
bugs.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>